ProyectoAnalizadorCVs/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
azure-ai-formrecognizer
azure-storage-blob
openai